Improved integration tests at the `main` level
